requirements.txt
requests
requests-oauthlib
python-dotenv
boto3